refactor(imports): unify UI component imports with alias and barrel files
- Create barrel files (index.ts) for all 11 pt-* components
- Update imports in quiz.container.ts to use @ui/<component> format
- Update imports in battle-card.ts to use barrel imports
- Update pt-chip and pt-type-chip internal imports
- Add ESLint rule to enforce @ui/<component> pattern
- Add guardrail documentation for UI import standards

Closes #89
