+++ src/app/features/quiz/components/battle-card/battle-card.ts
@@ -2,12 +2,12 @@ import { Component, Input, inject } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Pokemon } from '@domain/pokemon.schema';
 import { PokemonType, POKEMON_TYPES_MAP } from '@domain/type-chart';
-import { TypeChipComponent } from '@ui/pt-type-chip/pt-type-chip';
-import { AvatarComponent } from '@ui/pt-avatar/pt-avatar';
-import { IconComponent } from '@ui/pt-icon/pt-icon';
-import { StackComponent } from '@ui/pt-stack/pt-stack';
-import { SurfaceComponent } from '@ui/pt-surface/pt-surface';
-import { TextComponent } from '@ui/pt-text/pt-text';
+import { TypeChipComponent } from '@ui/pt-type-chip';
+import { AvatarComponent } from '@ui/pt-avatar';
+import { IconComponent } from '@ui/pt-icon';
+import { StackComponent } from '@ui/pt-stack';
+import { SurfaceComponent } from '@ui/pt-surface';
+import { TextComponent } from '@ui/pt-text';
 import { AssetPathService } from '@app/core/services/asset-path.service';
 
 /**
