+++ src/app/ui/pt-avatar/index.ts
@@ -1,0 +1,1 @@
+export { AvatarComponent } from './pt-avatar';
