+++ src/app/ui/pt-button/index.ts
@@ -1,0 +1,1 @@
+export { ButtonComponent } from './pt-button';
