+++ src/app/ui/pt-chip/index.ts
@@ -1,0 +1,3 @@
+export { ChipComponent } from './pt-chip';
+export type { ChipSize, ChipRounded } from './pt-chip.types';
+export { CHIP_SIZES, CHIP_ROUNDED } from './pt-chip.types';
