+++ src/app/ui/pt-grid/index.ts
@@ -1,0 +1,1 @@
+export { GridComponent } from './pt-grid';
