+++ src/app/ui/pt-heading/index.ts
@@ -1,0 +1,1 @@
+export { HeadingComponent } from './pt-heading';
