+++ src/app/ui/pt-icon/index.ts
@@ -1,0 +1,1 @@
+export { IconComponent } from './pt-icon';
