+++ src/app/ui/pt-spinner/index.ts
@@ -1,0 +1,1 @@
+export { SpinnerComponent } from './pt-spinner';
