+++ src/app/ui/pt-stack/index.ts
@@ -1,0 +1,1 @@
+export { StackComponent } from './pt-stack';
