+++ src/app/ui/pt-surface/index.ts
@@ -1,0 +1,1 @@
+export { SurfaceComponent } from './pt-surface';
