+++ src/app/ui/pt-text/index.ts
@@ -1,0 +1,1 @@
+export { TextComponent } from './pt-text';
