+++ src/app/ui/pt-type-chip/index.ts
@@ -1,0 +1,1 @@
+export { TypeChipComponent } from './pt-type-chip';
